+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [2.0.0] - 2024-11-29
 
 ### Changed
 
@@ -58,7 +58,7 @@
 
 - Increase version
 
-[Unreleased]: https://github.com/inlavigo/gg_json_hash/compare/1.1.4...HEAD
+[2.0.0]: https://github.com/inlavigo/gg_json_hash/compare/1.1.4...2.0.0
 [1.1.4]: https://github.com/inlavigo/gg_json_hash/compare/1.1.3...1.1.4
 [1.1.3]: https://github.com/inlavigo/gg_json_hash/compare/1.1.2...1.1.3
 [1.1.2]: https://github.com/inlavigo/gg_json_hash/compare/1.1.1...1.1.2